Ignore files already in store while computing segment file diff with primary
